requirements.txt
pillow==11.0.0
pygame==2.6.1
PySDL2==0.9.16
pyusb==1.2.1